Updated sha256 and paths in pyngl src [ci skip]
The PyNGL source code repository was recently restructured. Previously,
all source code was contained in a 'src' subdirectory, and the repo
included several directories of documentation and outdated test scripts.

The contents of the PyNGL repository now match what used to be in the
'src' subdirectory, so the build script no longer needs to 'cd src', and
the tests in meta.yaml should run from 'examples', not 'src/examples'.